refactor(test-filler): move fixture_output to shared
Move FixtureOutput from filler/ to shared/ and absorb fixture_naming.py.
This prepares for future build-tool plugins that also generate fixtures.
