+++ packages/testing/src/execution_testing/cli/pytest_commands/plugins/filler/__init__.py
@@ -1,6 +1,6 @@
 """A pytest plugin to fill tests and generate JSON fixtures."""
 
-from .fixture_output import FixtureOutput
+from ..shared.fixture_output import FixtureOutput
 
 __all__ = [
     "FixtureOutput",
