+++ packages/testing/src/execution_testing/cli/pytest_commands/plugins/filler/tests/test_generate_all_formats.py
@@ -2,7 +2,7 @@
 
 from typing import Any
 
-from execution_testing.cli.pytest_commands.plugins.filler.fixture_output import (  # noqa: E501
+from execution_testing.cli.pytest_commands.plugins.shared.fixture_output import (  # noqa: E501
     FixtureOutput,
 )
